Add command to create missing "images" directory
The "images" directory does not exist inside the "~/libvirt/" directory, so the command 
"mv mac_hdd_ng.img ~/libvirt/images" results in moving the file into "~/libvirt/" and renaming the file "mac_hdd_ng.img" "images".  

I suspect that is not the intention. So creating the "images" directory prior to the move will have the intended result of moving the "mac_hdd_ng.img" file into the directory "images" in the "libvirt" parent directory.